+++ tests/main_test.py
@@ -1,6 +1,7 @@
 """
 Test app & utilities.
 """
+
 import os
 
 from fastapi import FastAPI, Request, Response