+++ ycc_hull/db/context.py
@@ -1,6 +1,7 @@
 """
 Database context.
 """
+
 from typing import Any, Awaitable, TypeVar
 from collections.abc import Callable, Sequence