+++ ycc_hull/models/dtos.py
@@ -1,6 +1,7 @@
 """
 General API DTO classes.
 """
+
 from datetime import date
 
 from ycc_hull.db.entities import (